move code to internal/experimental and example to test
also update test that was no longer compiling
